Chess
This Chess was writed code in java. When I was created an code this project, I don´t know about data structure. This game is very simple, and only move the pieces on the board without restric rules of the chess, by example the movement of a queen, horse, tower, etc...